* origin/dev/5.0.0: (40 commits)
  add postman-combine-collection to project
  TM-16047 fixing for DST again
  TM--15939 - Fixed issue where 0's were being treated like nulls on asset show views
  TM-15483
  TM-16032 Avoid Task Spec in HTML render when it is null
  TM-16035: Removing the testingList references. The testing dropdown was removed because, this list was directly connected to Wrokflow, which is removed
  TM-15974 TM-15583 - Fixed issue with datetimes not syncing up between project list and project detail, also added completion vs. start date validation
  TM-15878 Action: Methods in the list are always displaying callEndpoint instead of the proper one, fixed issues with model
  TM-13237 It does optional day definitions in range filters
  TM-16034 - Fixed random typo on project menu
  TM-15974 - Small tweak to fix null deletion error
  TM-16047 Added a few different timezones
  TM-16032 Fixes from pull request
  TM-16019 - Minor refactors to event save/update
  TM-15507 Wrap the Tags column in the Asset List grids when it exceeds a single line, fix for when closing a modal
  TM-15949 - Fixed FE issues, NPE issue
  TM-15949 - Fixed FE issues, NPE issue
  TM-13237 Removed unused comment
  TM-16032 Fix cyclical results report
  TM-15983 - Fixed sort icon position on large name columns
  ...

# Conflicts:
#	src/main/groovy/com/tdsops/common/sql/SqlUtil.groovy

+++ grails-app/services/net/transitionmanager/project/ProjectService.groovy
@@ -1954,7 +1954,9 @@ class ProjectService implements ServiceMethods {
 		else {
 			if (!projectCommand.projectLogo) {
 				ProjectLogo projectLogo = ProjectLogo.findByProject(project)
-				projectLogo.delete()
+				if(projectLogo) {
+					projectLogo.delete()
+				}
 			}
 		}
